feat(tamanu): use the host's Seedling identity, elevating to read it
An operator with no Seedling identity of their own was told to generate one
and get it authorised, which is exactly the setup a freshly provisioned host
should not need: the Seedling package now authorises an identity for the host
itself. Fall back to that when the operator has none, and re-run elevated
when it (or the daemon's published identity, also root-only) is out of reach
unprivileged, the same way the host service manager path already elevates.

The operator's own identity still comes first where they have one, so the
daemon's record of who acted names them; acting under the host identity, name
the invoking operator to the daemon instead so the record is still a person.
A refusal on a host that carries an identity says that a privileged run would
be accepted, rather than leaving the operator to work that out.

Splits the module so the host-identity facts need no protocol client: alertd
observes them without pulling in the interface client.
